backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.39.8",
    "celery>=5.5.3",
    "dj-database-url>=3.0.1",
    "django>=5.2.4",
    "django-allauth>=65.10.0",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "django-filter>=25.1",
    "django-storages>=1.14.6",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-spectacular>=0.28.0",
    "factory-boy>=3.3.3",
    "gunicorn>=23.0.0",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.1",
    "pytest-django>=4.11.1",
    "python-decouple>=3.8",
    "redis>=6.2.0",
    "sentry-sdk>=2.33.0",
    "whitenoise>=6.9.0",
]
